requirements.txt
opencv-python
Pillow
numpy
matplotlib
requests
